requirements.txt
mcp>=1.13.0
langchain>=0.2.0
langgraph>=0.0.48
langchain-mcp-adapters>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0
